+++ shared/modules/persistence_module.go
@@ -3,6 +3,7 @@ package modules
 //go:generate mockgen -destination=./mocks/persistence_module_mock.go github.com/pokt-network/pocket/shared/modules PersistenceModule,PersistenceRWContext,PersistenceReadContext,PersistenceWriteContext
 
 import (
+	"github.com/pokt-network/pocket/persistence/indexer"
 	"github.com/pokt-network/pocket/persistence/kvstore"
 	"github.com/pokt-network/pocket/runtime/genesis"
 	coreTypes "github.com/pokt-network/pocket/shared/core/types"
@@ -24,7 +25,8 @@ type PersistenceModule interface {
 	GetBlockStore() kvstore.KVStore
 	NewWriteContext() PersistenceRWContext
 
-	// Indexer Queries
+	// Indexer operations
+	GetTxIndexer() indexer.TxIndexer
 	TransactionExists(transactionHash string) (bool, error)
 
 	// Debugging / development only
@@ -135,6 +137,12 @@ type PersistenceReadContext interface {
 	GetHeight() (int64, error) // Returns the height of the context
 	Release()                  // Releases the read context
 
+	// Version queries
+	GetVersionAtHeight(height int64) (string, error) // TODO: Implement this
+
+	// Supported Chains Queries
+	GetSupportedChains(height int64) ([]string, error) // TODO: Implement this
+
 	// CONSOLIDATE: BlockHash / AppHash / StateHash
 	// Block Queries
 	GetMaximumBlockHeight() (uint64, error)    // Returns the height of the latest block in the persistence layer
@@ -157,6 +165,7 @@ type PersistenceReadContext interface {
 	GetActor(actorType coreTypes.ActorType, address []byte, height int64) (*coreTypes.Actor, error)
 
 	// App Queries
+	GetApp(address []byte, height int64) (*coreTypes.Actor, error)
 	GetAllApps(height int64) ([]*coreTypes.Actor, error)
 	GetAppExists(address []byte, height int64) (exists bool, err error)
 	GetAppStakeAmount(height int64, address []byte) (string, error)
@@ -166,6 +175,7 @@ type PersistenceReadContext interface {
 	GetAppOutputAddress(operator []byte, height int64) (output []byte, err error)
 
 	// Servicer Queries
+	GetServicer(address []byte, height int64) (*coreTypes.Actor, error)
 	GetAllServicers(height int64) ([]*coreTypes.Actor, error)
 	GetServicerExists(address []byte, height int64) (exists bool, err error)
 	GetServicerStakeAmount(height int64, address []byte) (string, error)
@@ -176,6 +186,7 @@ type PersistenceReadContext interface {
 	GetServicerCount(chain string, height int64) (int, error)
 
 	// Fisherman Queries
+	GetFisherman(address []byte, height int64) (*coreTypes.Actor, error)
 	GetAllFishermen(height int64) ([]*coreTypes.Actor, error)
 	GetFishermanExists(address []byte, height int64) (exists bool, err error)
 	GetFishermanStakeAmount(height int64, address []byte) (string, error)
@@ -185,6 +196,7 @@ type PersistenceReadContext interface {
 	GetFishermanOutputAddress(operator []byte, height int64) (output []byte, err error)
 
 	// Validator Queries
+	GetValidator(address []byte, height int64) (*coreTypes.Actor, error)
 	GetAllValidators(height int64) ([]*coreTypes.Actor, error)
 	GetValidatorExists(address []byte, height int64) (exists bool, err error)
 	GetValidatorStakeAmount(height int64, address []byte) (string, error)