+++ utility/types/message_test.go
@@ -40,15 +40,15 @@ func TestMessage_ChangeParameter_ValidateBasic(t *testing.T) {
 
 	msgMissingOwner := proto.Clone(&msg).(*MessageChangeParameter)
 	msgMissingOwner.Owner = nil
-	require.Equal(t, ErrEmptyAddress().Code(), msgMissingOwner.ValidateBasic().Code())
+	require.Equal(t, coreTypes.ErrEmptyAddress().Code(), msgMissingOwner.ValidateBasic().Code())
 
 	msgMissingParamKey := proto.Clone(&msg).(*MessageChangeParameter)
 	msgMissingParamKey.ParameterKey = ""
-	require.Equal(t, ErrEmptyParamKey().Code(), msgMissingParamKey.ValidateBasic().Code())
+	require.Equal(t, coreTypes.ErrEmptyParamKey().Code(), msgMissingParamKey.ValidateBasic().Code())
 
 	msgMissingParamValue := proto.Clone(&msg).(*MessageChangeParameter)
 	msgMissingParamValue.ParameterValue = nil
-	require.Equal(t, ErrEmptyParamValue().Code(), msgMissingParamValue.ValidateBasic().Code())
+	require.Equal(t, coreTypes.ErrEmptyParamValue().Code(), msgMissingParamValue.ValidateBasic().Code())
 }
 
 func TestMessage_EditStake_ValidateBasic(t *testing.T) {
@@ -67,33 +67,33 @@ func TestMessage_EditStake_ValidateBasic(t *testing.T) {
 	msgMissingAmount := proto.Clone(&msg).(*MessageEditStake)
 	msgMissingAmount.Amount = ""
 	er := msgMissingAmount.ValidateBasic()
-	require.Equal(t, ErrEmptyAmount().Code(), er.Code())
+	require.Equal(t, coreTypes.ErrEmptyAmount().Code(), er.Code())
 
 	msgInvalidAmount := proto.Clone(&msg).(*MessageEditStake)
 	msgInvalidAmount.Amount = "sdk"
 	er = msgInvalidAmount.ValidateBasic()
-	require.Equal(t, ErrStringToBigInt(er).Code(), er.Code())
+	require.Equal(t, coreTypes.ErrStringToBigInt(er).Code(), er.Code())
 
 	msgEmptyAddress := proto.Clone(&msg).(*MessageEditStake)
 	msgEmptyAddress.Address = nil
 	er = msgEmptyAddress.ValidateBasic()
-	require.Equal(t, ErrEmptyAddress().Code(), er.Code())
+	require.Equal(t, coreTypes.ErrEmptyAddress().Code(), er.Code())
 
 	msgInvalidAddress := proto.Clone(&msg).(*MessageEditStake)
 	msgInvalidAddress.Address = []byte("badAddr")
 	er = msgInvalidAddress.ValidateBasic()
-	expectedErr := ErrInvalidAddressLen(crypto.ErrInvalidAddressLen(defaultUnusedLength))
+	expectedErr := coreTypes.ErrInvalidAddressLen(crypto.ErrInvalidAddressLen(defaultUnusedLength))
 	require.Equal(t, expectedErr.Code(), er.Code())
 
 	msgEmptyRelayChains := proto.Clone(&msg).(*MessageEditStake)
 	msgEmptyRelayChains.Chains = nil
 	er = msgEmptyRelayChains.ValidateBasic()
-	require.Equal(t, ErrEmptyRelayChains().Code(), er.Code())
+	require.Equal(t, coreTypes.ErrEmptyRelayChains().Code(), er.Code())
 
 	msgInvalidRelayChains := proto.Clone(&msg).(*MessageEditStake)
 	msgInvalidRelayChains.Chains = []string{"notAValidRelayChain"}
 	er = msgInvalidRelayChains.ValidateBasic()
-	expectedErr = ErrInvalidRelayChainLength(0, relayChainLength)
+	expectedErr = coreTypes.ErrInvalidRelayChainLength(0, relayChainLength)
 	require.Equal(t, expectedErr.Code(), er.Code())
 }
 
@@ -115,22 +115,22 @@ func TestMessage_Send_ValidateBasic(t *testing.T) {
 	msgMissingAddress := proto.Clone(&msg).(*MessageSend)
 	msgMissingAddress.FromAddress = nil
 	er = msgMissingAddress.ValidateBasic()
-	require.Equal(t, ErrEmptyAddress().Code(), er.Code())
+	require.Equal(t, coreTypes.ErrEmptyAddress().Code(), er.Code())
 
 	msgMissingToAddress := proto.Clone(&msg).(*MessageSend)
 	msgMissingToAddress.ToAddress = nil
 	er = msgMissingToAddress.ValidateBasic()
-	require.Equal(t, ErrEmptyAddress().Code(), er.Code())
+	require.Equal(t, coreTypes.ErrEmptyAddress().Code(), er.Code())
 
 	msgMissingAmount := proto.Clone(&msg).(*MessageSend)
 	msgMissingAmount.Amount = ""
 	er = msgMissingAmount.ValidateBasic()
-	require.Equal(t, ErrEmptyAmount().Code(), er.Code())
+	require.Equal(t, coreTypes.ErrEmptyAmount().Code(), er.Code())
 
 	msgInvalidAmount := proto.Clone(&msg).(*MessageSend)
 	msgInvalidAmount.Amount = ""
 	er = msgInvalidAmount.ValidateBasic()
-	require.Equal(t, ErrEmptyAmount().Code(), er.Code())
+	require.Equal(t, coreTypes.ErrEmptyAmount().Code(), er.Code())
 }
 
 func TestMessage_Stake_ValidateBasic(t *testing.T) {
@@ -151,22 +151,22 @@ func TestMessage_Stake_ValidateBasic(t *testing.T) {
 	msgEmptyPubKey := proto.Clone(&msg).(*MessageStake)
 	msgEmptyPubKey.PublicKey = nil
 	er = msgEmptyPubKey.ValidateBasic()
-	require.Equal(t, ErrEmptyPublicKey().Code(), er.Code())
+	require.Equal(t, coreTypes.ErrEmptyPublicKey().Code(), er.Code())
 
 	msgEmptyChains := proto.Clone(&msg).(*MessageStake)
 	msgEmptyChains.Chains = nil
 	er = msgEmptyChains.ValidateBasic()
-	require.Equal(t, ErrEmptyRelayChains().Code(), er.Code())
+	require.Equal(t, coreTypes.ErrEmptyRelayChains().Code(), er.Code())
 
 	msgEmptyAmount := proto.Clone(&msg).(*MessageStake)
 	msgEmptyAmount.Amount = ""
 	er = msgEmptyAmount.ValidateBasic()
-	require.Equal(t, ErrEmptyAmount().Code(), er.Code())
+	require.Equal(t, coreTypes.ErrEmptyAmount().Code(), er.Code())
 
 	msgEmptyOutputAddress := proto.Clone(&msg).(*MessageStake)
 	msgEmptyOutputAddress.OutputAddress = nil
 	er = msgEmptyOutputAddress.ValidateBasic()
-	require.Equal(t, ErrNilOutputAddress().Code(), er.Code())
+	require.Equal(t, coreTypes.ErrNilOutputAddress().Code(), er.Code())
 }
 
 func TestMessage_Unstake_ValidateBasic(t *testing.T) {
@@ -182,7 +182,7 @@ func TestMessage_Unstake_ValidateBasic(t *testing.T) {
 	msgMissingAddress := proto.Clone(&msg).(*MessageUnstake)
 	msgMissingAddress.Address = nil
 	er = msgMissingAddress.ValidateBasic()
-	require.Equal(t, ErrEmptyAddress().Code(), er.Code())
+	require.Equal(t, coreTypes.ErrEmptyAddress().Code(), er.Code())
 }
 
 func TestMessage_Unpause_ValidateBasic(t *testing.T) {
@@ -198,5 +198,5 @@ func TestMessage_Unpause_ValidateBasic(t *testing.T) {
 	msgMissingAddress := proto.Clone(&msg).(*MessageUnpause)
 	msgMissingAddress.Address = nil
 	er = msgMissingAddress.ValidateBasic()
-	require.Equal(t, ErrEmptyAddress().Code(), er.Code())
+	require.Equal(t, coreTypes.ErrEmptyAddress().Code(), er.Code())
 }